M    test/edu/ucla/sspace/matrix/MatrixIOtests.java

- Updated test to correct svdlibc format

M    src/edu/ucla/sspace/matrix/MatrixIO.java

- Fixed svdlibc sparse bug in readMatrixArray() that used wrong precision for
  values

- Added method for getting an iterator over the values in a matrix file

A    src/edu/ucla/sspace/matrix/SvdlibcSparseBinaryFileIterator.java
A    src/edu/ucla/sspace/matrix/MatlabSparseFileIterator.java

- Added two implementions for iterating over the values in their respective
  matrix file formats.

A    src/edu/ucla/sspace/matrix/MatrixEntry.java
A    src/edu/ucla/sspace/matrix/SimpleEntry.java

- Added interface for abstracting matrix values and a package private
  implementation used by the iterators

A    test/edu/ucla/sspace/matrix/SvdlibcSparseBinaryFileIteratorTests.java
A    test/edu/ucla/sspace/matrix/MatlabSparseFileIteratorTests.java

- Added tests for the iterators

M    src/edu/ucla/sspace/matrix/LogEntropyTransform.java

- Clarified in Javadoc which paper the implementation is based on

- Fixed bug where the matlab-adjusted indices were not corrected when writing
  the final indices back to the tranformed matrix file

- Clarified internal comments

M    build.xml

- Added unit tests for matrix iterators

+++ trunk/sspace/test/edu/ucla/sspace/matrix/MatrixIOtests.java
@@ -80,21 +80,21 @@ public class MatrixIOtests {
 
         pw.writeInt(2);
         pw.writeInt(0);
-        pw.writeDouble(2.3);
+        pw.writeFloat(2.3f);
         pw.writeInt(2);
-        pw.writeDouble(3.8);
+        pw.writeFloat(3.8f);
 
         pw.writeInt(1);
         pw.writeInt(1);
-        pw.writeDouble(1.3);
+        pw.writeFloat(1.3f);
 
         pw.writeInt(3);
         pw.writeInt(0);
-        pw.writeDouble(4.2);
+        pw.writeFloat(4.2f);
         pw.writeInt(1);
-        pw.writeDouble(2.2);
+        pw.writeFloat(2.2f);
         pw.writeInt(2);
-        pw.writeDouble(0.5);
+        pw.writeFloat(0.5f);
 
         pw.close();
         return f;
